core: fiber inhertance refactor start
