+++ src/core/fiber_container.cpp
@@ -4,6 +4,7 @@
 
 #include <body.hpp>
 #include <fiber.hpp>
+#include <fiber_container.hpp>
 #include <kernels.hpp>
 #include <periphery.hpp>
 #include <system.hpp>
